requirements.txt
absl-py==0.7.1
astor==0.8.0
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
Flask==1.0.3
gast==0.2.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imutils==0.5.2
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.8.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.4
wincertstore==0.2
wrapt==1.11.1
